Reduce polling in main Game
Polling was previously set to every 3 seconds for fetching all phrases and checking for updates to the main game (primarily to check for round updates). After a test run, actual usage indicates every 5 seconds should be sufficient, but eventually this should be done with websockets or something more efficient than polling.